Refactor for direct usage of library in other libraries. Previously FindDeps was not accessible due to main package
