+++ .gitignore
@@ -1,1 +1,2 @@
 /deplist
+.idea
