--- testdata/src/a/a.go
@@ -1,7 +1,0 @@
-package main
-
-import "b"
-
-func main() {
-	b.B()
-}
