--- testdata/src/a/a_test.go
@@ -1,5 +1,0 @@
-package main
-
-import "d"
-
-const D = d.D
