--- testdata/src/b/b.go
@@ -1,7 +1,0 @@
-package b
-
-import "c"
-
-func B() {
-	c.C()
-}
