--- testdata/src/d/d.go
@@ -1,3 +1,0 @@
-package d
-
-const D = true
