--- testdata/src/e/e.go
@@ -1,7 +1,0 @@
-package main
-
-import "v0"
-
-func main() {
-	v0.V()
-}
